Almost good version, some errors still to be fixed though.

+++ src/code/scripts/shell/sort-reiner-mirror.sh
@@ -6,18 +6,19 @@ Usage : "`basename $0`" <Reiner's mirror> <target repository for sorted content>
 MV=/bin/mv
 RM=/bin/rm
 MKDIR=/bin/mkdir
+RMDIR=/bin/rmdir
 
-# Recommended usage :
+ERROR_OUTPUT="2>/dev/null"
+
+# Recommended usage :
 # 1. Perform a very slow wget (to avoid disturbing anybody) on Reiner's site :
-# wget --limit-rate=10k --wait=5 --random-wait --retry-connrefused --recursive --level=10 http://reinerstileset.4players.de
-# 2. Archive it in ~/Archive/Content/<date>-reiners-tileset-mirror.tar.bz2
+# wget --limit-rate=10k --wait=5 --random-wait --retry-connrefused --recursive --level=10 --reject 'M_*.zip' --exclude-directories=forum http://reinerstileset.4players.de 
+# 2. Archive it in ~/Archive/Content/<date>-reiners-tileset-mirror.tar.bz2
 # 3. Extract a copy of this archive to be used here as Reiner's mirror 
 # (its content will be removed)
 
-
-if [ $# -neq 2 ] ; then
+if [ ! $# -eq 2 ] ; then
 	echo "
-	
 	Error, two arguments expected. $USAGE
 	
 	" 1>&2
@@ -28,7 +29,6 @@ MIRROR_ROOT="$1"
 
 if [ ! -d "${MIRROR_ROOT}" ] ; then
 	echo "
-	
 	Error, no directory named <${MIRROR_ROOT}> exists for mirror. $USAGE
 	
 	" 1>&2
@@ -37,38 +37,32 @@ fi
 
 TARGET_ROOT="$2"
 
+# Absolute path only :
+first_char=`echo ${TARGET_ROOT} | cut -b 1`
+if [ ${first_char} != "/" ] ; then
+	TARGET_ROOT=`pwd`/${TARGET_ROOT}
+fi
+
+
 if [ ! -d "${TARGET_ROOT}" ] ; then
 	echo "
-	
 	Error, no directory named <${TARGET_ROOT}> exists for target repository. $USAGE
 	
 	" 1>&2
 	exit 3
 fi
 	
-echo "
 
-	Will sort Reiner's tileset mirror now...
-	
-	"
+# Look in Ceylan sources :
+CEYLAN_LOANI_OFFSET="../../../../../../../ceylan/Ceylan/trunk/src/code/scripts/shell/"
+
+# Anticipated checkings :
+CORRECTER=`pwd`"/"`dirname $0`"/${CEYLAN_LOANI_OFFSET}/correctPathsInTree.sh"
 
 cd ${MIRROR_ROOT}
 
-if [ ! -f "englisch.htm" ] ; then
-	echo "
-	
-	Error, not in Reiner mirror root ?
-	" 1>&2
-	exit 4
-fi
-
-
-# Anticipated checkings :
-CORRECTER=`dirname $0`/correctPathsInTree.sh
-
 if [ ! -x "${CORRECTER}" ] ; then
 	echo "
-	
 	Error, no executable correcter script found (searched ${CORRECTER}).
 	
 	" 1>&2
@@ -77,34 +71,55 @@ if [ ! -x "${CORRECTER}" ] ; then
 fi
 
 
+if [ ! -f "englisch.htm" ] ; then
+	echo "
+	Error, not in Reiner mirror root ? (here is "`pwd`")
+	" 1>&2
+	exit 4
+fi
+
+
+echo "
+
+	Will sort Reiner's tileset mirror now...
+	
+	"
+
 echo "  + removing non-wanted content"
 
 
 # M_* are MMF files :
 find . \( -name '*.exe' -o -name '*.htm*' -o -name '*.wmv' -o -name 'M*.zip'  \) -exec ${RM} -f '{}' ';'
 
-echo "  + normalizing paths"
-${CORRECTER} .
+# Supress non-wanted directories :
+${RM} -rf forum *uvmappingtut* screenshots buttons handshaped_3d_tree_htm_files ext mmfabsbeg
 
-# 'T ' became 'T-', we want 'T_' for uniformity :
+# Supress non-wanted files :
+${RM} AAspritesharpborder1-4.zip bass.zip *'enemie ai'*.zip gesleiste.zip isotiletuteng.zip *Tutorial* *Sbonestut.zip *jumpengine* 2>/dev/null
+
+
+echo "  + normalizing paths"
+
+echo "    - correcting paths"
+${CORRECTER} . | grep -v 'left unchanged'
+
+echo "    - correcting underscores #1"
+# 'T ' became 'T-', we want 'T_' for uniformity :
 for f in `find . -name 'T-*.zip'` ; do
 	corrected=`echo $f | tr '-' '_'`
 	${MV} -f $f $corrected	
 done 
 
-# No 'Tgrundvari.zip', we want 'T_grundvari.zip' :
+echo "    - correcting underscores #2"
+# No 'Tgrundvari.zip', we want 'T_grundvari.zip' :
 for f in `find . -name 'T*.zip'` ; do
-	if [ `echo $f | cut -b 4` != '_' ] ; then
+	fourth_char=`echo $f | cut -b 4`
+	if [ "${fourth_char}" != "_" ] ; then
 		corrected=`echo $f | sed 's|^./T|./T_|1'`
 		${MV} -f $f $corrected
 	fi
 done 
 
-# Supress non-wanted directories :
-${RM} -rf forum uvmappingtut_-htm_files screenshots buttons handshaped_3d_tree_htm_files ext mmfabsbeg
-
-# Supress non-wanted files :
-${RM} AAspritesharpborder1-4.zip bass.zip enemie-ai-for-more-enemies-final-.zip gesleiste.zip isotiletuteng.zip T_Sbonestut.zip *-Tutorial-* *jumpengine*
 
 # Rename for convenience some files :
 ${MV} 8Dshooting-fish.zip T_shooting_fish.zip
@@ -132,7 +147,7 @@ ${MV} T_john_doe.zip T_john_doe_santa.zip
 
 ##########################################################################
 
-# Directory definitions (sorted alphabetically) :
+# Directory definitions (sorted alphabetically) :
 
 # In root :
 
@@ -143,7 +158,7 @@ SOUND_ROOT=${TARGET_ROOT}/Sounds
 ${MKDIR} ${SOUND_ROOT}
 
 
-# In Graphics :
+# In Graphics :
 
 CREATURES=${GRAPHICAL_ROOT}/Creatures
 ${MKDIR} ${CREATURES}
@@ -407,7 +422,7 @@ ${MKDIR} ${WEAPONS}
 
 
 
-# In Vehicles :
+# In Vehicles :
 
 BOATS=${VEHICLES}/Boats
 ${MKDIR} ${BOATS}
@@ -425,7 +440,7 @@ ARMORED=${VEHICLES}/Armored
 ${MKDIR} ${ARMORED}
 
 
-# In Environments :
+# In Environments :
 
 CASTLES=${ENVIRONMENTS}/Castles
 ${MKDIR} ${CASTLES}
@@ -468,7 +483,7 @@ OUTSIDE=${WALLS}/Outside
 ${MKDIR} ${OUTSIDE}
 
 
-# In Sprites :
+# In Sprites :
 
 GUI=${SPRITES}/GUI
 ${MKDIR} ${GUI}
@@ -479,97 +494,100 @@ ${MKDIR} ${GUI}
 
 # Transfer rules :
 
-${MV} sounds/*   ${SOUND_ROOT}
-${MV} sprites/*  ${SPRITES}
-${MV} *menu* *button* *gui* *butscr* *levelup* *pointrunes* ${GUI}
-${MV} tilesets/* ${TILESETS}
+${MV} sounds/*   ${SOUND_ROOT} 
+${MV} sprites/*  ${SPRITES} 
+${MV} tilesets/* ${TILESETS} 
+${RMDIR} sounds sprites tilesets
+
+${MV} *menu* *button* *gui* *butscr* *levelup* *pointrunes* ${GUI}  
 
 
 # Creatures :
 
-${MV} *groggy* ${CREATURES}
+${MV} *groggy* ${CREATURES} 
 
 
-# Humans :
+# Humans :
 
-${MV} *flying_soul*    ${HUMANS}
+${MV} *flying_soul*    ${HUMANS} 
 
-${MV} *archer*          ${ARCHERS}
-${MV} *baker.*          ${BAKERS}
-${MV} *barbar*          ${BARBARIANS}
-${MV} *beekeeper*       ${BEEKEEPERS}
-${MV} *boy* *girl*      ${CHILDREN}
-${MV} *builder*         ${BUILDERS}
-${MV} *carrier*         ${CARRIERS}
-${MV} *caveman*         ${CAVEMEN}
-${MV} *doc.*            ${DOCTORS}
-${MV} *farmer*          ${FARMERS}
-${MV} *fisherman*       ${FISHERMEN}
-${MV} *hunter*          ${HUNTERS}
-${MV} *knight*          ${KNIGHTS}
-${MV} *konig* *prince*  ${ROYALTY}
-${MV} *lumberjack*      ${LUMBERJACKS}
-${MV} *mage* *magier*   ${MAGES}
-${MV} *miller*          ${MILLERS}
-${MV} *miner*           ${MINERS}
-${MV} *monk*            ${MONKS}
-${MV} *ninja*           ${NINJAS}
-${MV} *pirsa* *nobeard* ${PIRATES}
-${MV} *professor*       ${PROFESSORS}
-${MV} *_smith.*         ${SMITHS}
-${MV} *soldier*         ${SOLDIERS}
-${MV} *swordsman*       ${SWORDSMEN}
-${MV} *thief*           ${THIEVES}
+${MV} *archer*          ${ARCHERS} 
+${MV} *baker.*          ${BAKERS} 
+${MV} *barbar*          ${BARBARIANS} 
+${MV} *beekeeper*       ${BEEKEEPERS} 
+${MV} *boy* *girl*      ${CHILDREN} 
+${MV} *builder*         ${BUILDERS} 
+${MV} *carrier*         ${CARRIERS} 
+${MV} *caveman*         ${CAVEMEN} 
+${MV} *doc.*            ${DOCTORS} 
+${MV} *farmer*          ${FARMERS} 
+${MV} *fisherman*       ${FISHERMEN} 
+${MV} *hunter*          ${HUNTERS} 
+${MV} *knight*          ${KNIGHTS} 
+${MV} *konig* *prince*  ${ROYALTY} 
+${MV} *lumberjack*      ${LUMBERJACKS} 
+${MV} *mage* *magier*   ${MAGES} 
+${MV} *miller*          ${MILLERS} 
+${MV} *miner*           ${MINERS} 
+${MV} *monk*            ${MONKS} 
+${MV} *ninja*           ${NINJAS} 
+${MV} *pirsa* *nobeard* ${PIRATES} 
+${MV} *professor*       ${PROFESSORS} 
+${MV} *_smith.*         ${SMITHS} 
+${MV} *soldier*         ${SOLDIERS} 
+${MV} *swordsman*       ${SWORDSMEN} 
+${MV} *thief*           ${THIEVES} 
 
 
-# Named characters :
+# Named characters :
 
-${MV} *anna*  ${ANNA}
-${MV} *arno*  ${ARNO}
-${MV} *billy* ${BILLY}
-${MV} *bjorn* ${BJORN}
-${MV} *burra* ${BURRA}
-${MV} *elsa*  ${ELSA}
-${MV} *freya* ${FREYA}
-${MV} *horst* ${HORST}
-${MV} *john_doe_santa* ${JOHN}
-${MV} *katie* ${KATIE}
-${MV} *lilly* ${LILLY}
-${MV} *luigi* ${LUIGI}
-${MV} *luzia* ${LUZIA}
-${MV} *mrs_ruin* ${MRSRUIN}
-${MV} *rosalila* ${ROSALILA}
-${MV} *stan*  ${STAN}
-${MV} *vlad*  ${VLAD}
-${MV} *willy* ${WILLY}
+${MV} *anna*  ${ANNA} 
+${MV} *arno*  ${ARNO} 
+${MV} *billy* ${BILLY} 
+${MV} *bjorn* ${BJORN} 
+${MV} *burra* ${BURRA} 
+${MV} *elsa*  ${ELSA} 
+${MV} *freya* ${FREYA} 
+${MV} *horst* ${HORST} 
+${MV} *john_doe_santa* ${JOHN} 
+${MV} *katie* ${KATIE} 
+${MV} *lilly* ${LILLY} 
+${MV} *luigi* ${LUIGI} 
+${MV} *luzia* ${LUZIA} 
+${MV} *mrs_ruin* ${MRSRUIN} 
+${MV} *rosalila* ${ROSALILA} 
+${MV} *stan*  ${STAN} 
+${MV} *vlad*  ${VLAD} 
+${MV} *willy* ${WILLY} 
 
 # Animals :
 
-${MV} *donkey* *bat* *chicken* *cow* *crow* *deer* *_mouse.* *rabbit* *roebuck* *rooster* *schwein* ${ANIMALS}
+${MV} *donkey* *bat.* *chicken* *cow* *crow* *deer* *_mouse.* *rabbit* *roebuck* *rooster* *schwein* ${ANIMALS} 
 
-${MV} *dragonfly* *butterfly* ${INSECTS}
-${MV} *fish*   ${FISH}
-${MV} *wasp*   ${WASPS}
-${MV} *spider* ${SPIDERS}
-${MV} *wolf*   ${WOLVES}
-${MV} *lion*   ${LIONS}
-${MV} *sheep*  ${SHEEP}
+${MV} *dragonfly* *butterfly* ${INSECTS} 
+${MV} *fish*   ${FISH} 
+${MV} *wasp*   ${WASPS} 
+${MV} *spider* ${SPIDERS} 
+${MV} *wolf*   ${WOLVES} 
+${MV} *lion*   ${LIONS} 
+${MV} *sheep*  ${SHEEP} 
 
 # Monsters :
 
-${MV} *devil* *ogre* *fairy* *skull* *ghost* *theorc* *mummy* *mushy* *puff* *Sworm* ${MONSTERS}
+${MV} *doteater*     ${DOTEATERS} 
 
-${MV} *dragon*       ${DRAGONS}
-${MV} *dwarf*        ${DWARVES}
-${MV} *gnome*        ${GNOMES}
-${MV} *troll*        ${TROLLS}
-${MV} *zombie*       ${ZOMBIES}
-${MV} *ratz*         ${RATS}
-${MV} *dino* *pteri* ${DINOSAURS}
-${MV} *doteater*     ${DOTEATERS}
-${MV} *mister_eye*   ${CYCLOPS}
-${MV} *crocy* *swampthing* ${REPTILES}
-${MV} *skel* *mister_death* ${SKELETONS}
+${MV} *devil* *ogre* *fairy* *skull* *ghost* *theorc* *mummy* *mushy* *puff* *Sworm* ${MONSTERS} 
+
+${MV} *dragon*       ${DRAGONS} 
+${MV} *dwarf*        ${DWARVES} 
+${MV} *gnome*        ${GNOMES} 
+${MV} *troll*        ${TROLLS} 
+${MV} *zombie*       ${ZOMBIES} 
+${MV} *ratz*         ${RATS} 
+${MV} *dino* *pteri* ${DINOSAURS} 
+${MV} *mister_eye*   ${CYCLOPS} 
+${MV} *crocy* *swampthing* ${REPTILES} 
+${MV} *skel* *mister_death* ${SKELETONS} 
 
 
 
@@ -577,75 +595,75 @@ ${MV} *skel* *mister_death* ${SKELETONS}
 
 
 # Weapons :
-${MV} *cannon* *arrow* *tnt* *ammo* *pickup_weapon* *rotating_spike* *spear_catapult* ${WEAPONS}
+${MV} *cannon* *arrow* *tnt* *ammo* *pickup_weapon* *rotating_spike* *spear_catapult* ${WEAPONS} 
  
 # Mechanisms : 
-${MV} *mecha* *lever* *rotating_switch* ${MECHANISMS}
+${MV} *mecha* *lever* *rotating_switch* ${MECHANISMS} 
 
 # Traps :
-${MV} *trap* *mines* *spikefield* *spiketrench* ${TRAPS}
+${MV} *trap* *mines* *spikefield* *spiketrench* ${TRAPS} 
 
 # Intangible :
-${MV} *flamewall* *aura* *feuer* *bubble* *leafstorm* *magic_ani* ${INTANGIBLE}
+${MV} *flamewall* *aura* *feuer* *bubble* *leafstorm* *magic_ani* ${INTANGIBLE} 
 
 # Pieces of furniture :
-${MV} *beds* *carpet* *chest* *door* *furniture* *hedge_column* *keycolumn* *kitchen* *library* *pool_* *Sdrahomix* *shelfes* *shops_2* *throne_room* ${FURNITURE}
+${MV} *beds* *carpet* *chest* *door* *furniture* *hedge_column* *keycolumn* *kitchen* *library* *pool_* *Sdrahomix* *shelfes* *shops_2* *throne_room* ${FURNITURE} 
 
-${MV} *doormask* *wallpictures* *objects* *fields* *box* *item* *scare* *things* *gym* *lantern* *pickup_armor* *pickup_strategy* *_barrel* *rolling_stone* *sarkophag* *soulstones* *stoneheap* ${OBJECTS}
+${MV} *doormask* *wallpictures* *objects* *fields* *box* *item* *scare* *things* *gym* *lantern* *pickup_armor* *pickup_strategy* *_barrel* *rolling_stone* *sarkophag* *soulstones* *stoneheap* ${OBJECTS} 
 
 
 # Vehicles :
 
 # Boats :
-${MV} *schiff* *boot* *grokanbo* *grpirschi* *klpirschi* ${BOATS}
+${MV} *schiff* *boot* *grokanbo* *grpirschi* *klpirschi* ${BOATS} 
 
 # Spaceships :
-${MV} *spaceship* ${SPACESHIPS}
+${MV} *spaceship* ${SPACESHIPS} 
 
-# Cars :
-${MV} *rallye* ${CARS}
+# Cars :
+${MV} *rallye* ${CARS} 
 
 # Trains :
-${MV} *waggon* *lore.* *steam_engine* *woodwag* ${TRAINS}
+${MV} *waggon* *lore.* *steam_engine* *woodwag* ${TRAINS} 
 
 # Armored :
-${MV} *tank* ${ARMORED}
+${MV} *tank* ${ARMORED} 
 
-# Food :
-${MV} *meat* ${FOOD}
+# Food :
+${MV} *meat* ${FOOD} 
 
 
 
 # Environments :
 
-${MV} *castle* ${CASTLES}
+${MV} *castle* ${CASTLES} 
 
-${MV} *hills* ${HILLS}
+${MV} *hill* ${HILLS} 
 
-${MV} *set* *anitown* *beautiful_town* *casino* *cavestuff* *pipeline* *Sdraho* *Sdungeon* *thingmachine_all* ${SETS}
+${MV} *set* *anitown* *beautiful_town* *casino* *cavestuff* *pipeline* *Sdraho* *Sdungeon* *thingmachine_all* *actrumb* ${SETS} 
 
 
-# Walls :
+# Walls :
 
-${MV} *inside_wall* *quarry* *shops_1* *ug_fence* ${INSIDE}
-${MV} *wall* *parapet* *msw* *palisade* *woodenfence* ${OUTSIDE}
+${MV} *inside_wall* *quarry* *shops_1* *ug_fence* ${INSIDE} 
+${MV} *wall* *parapet* *msw* *palisade* *woodenfence* ${OUTSIDE} 
 
 # Stairs :
-${MV} *stairs* ${STAIRS}
+${MV} *stairs* ${STAIRS} 
 
 # Buildings :
 
-${MV} *house* *mill* *bakery* *barracks* *chapel* *crabuil* *building* *lodge* *hotel* *hut* *market_place* *_mine* *palace* *pottery* *_pub* *railroad* *ruins* *shipyard* *temple* *smithy* *_stable.* *triumph_arc* ${BUILDINGS}
+${MV} *house* *mill* *bakery* *barracks* *chapel* *crabuil* *building* *lodge* *hotel* *hut* *market_place* *_mine* *palace* *pottery* *_pub* *railroad* *ruins* *shipyard* *temple* *smithy* *_stable.* *triumph_arc* ${BUILDINGS}  
 
 
 # Plants :
 
-${MV} *tree* *bush* *growing_fir* ${TREES}
-${MV} *growing* *reef* ${PLANTS}
+${MV} *tree* *bush* *growing_fir* ${TREES} 
+${MV} *growing* *reef* ${PLANTS} 
 
 
-# Grounds :
-${MV} *earth* *wall* *beach* *cobble* *creek* *sand* *ground* *stonepath* *slope* *grund* *mountain* *pathway* *pfad* *quai* *_road_* *_rocks_* *tiled_floor* *weg_mit_randstein* ${GROUNDS}
+# Grounds :
+${MV} *earth* *wall* *beach* *cobble* *creek* *sand* *ground* *stonepath* *slope* *grund* *mountain* *pathway* *pfad* *quai* *_road_* *_rocks_* *tiled_floor* *weg_mit_randstein* *ani_water* ${GROUNDS} 
 
 
 echo "End of mirror sorting."
